+++ Neo4j.Driver/Neo4j.Driver/Internal/Auth/RotatingClientCertificateProvider.cs
@@ -16,6 +16,7 @@
 using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 using System.Threading.Tasks;
+using Neo4j.Driver.Preview.Auth;
 
 namespace Neo4j.Driver.Internal.Auth;
