+++ Neo4j.Driver/Neo4j.Driver/Internal/Auth/StaticClientCertificateProvider.cs
@@ -15,6 +15,7 @@
 
 using System.Security.Cryptography.X509Certificates;
 using System.Threading.Tasks;
+using Neo4j.Driver.Preview.Auth;
 
 namespace Neo4j.Driver.Internal.Auth;
