+++ Neo4j.Driver/Neo4j.Driver/Public/Config.cs
@@ -22,6 +22,7 @@ using Neo4j.Driver.Internal;
 using Neo4j.Driver.Internal.IO;
 using Neo4j.Driver.Internal.Logging;
 using Neo4j.Driver.Internal.Util;
+using Neo4j.Driver.Preview.Auth;
 
 namespace Neo4j.Driver;
